shared_interface/requirements.txt
flask==3.0.0
celery==5.3.5 
SQLAlchemy==2.0.23
mysqlclient==2.2.0
PyJWT==2.8.0